backend/pyproject.toml
[project]
name = "plaque-risk-explorer"
version = "0.1.0"
description = "Association of Clinical Factors and Plaque Morphology with Adverse Cardiovascular Outcomes"
authors = [
    { name = "Viacheslav" }
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "autogluon>=1.5.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pre-commit>=4.5.1",
    "pytest>=8.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ml", "src/pipelines", "src/plaque_risk_explorer"]

[tool.mypy]
python_version = "3.13"
exclude = ["venv", ".venv", "tests"]
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_untyped_calls = false
explicit_package_bases = true

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["plaque_risk_explorer", "ml", "pipelines"]
